requirements.txt
pymongo~=3.11.0
Flask~=1.1.2
python-dotenv~=0.15.0
pyopenssl